remove QPainterPath

+++ src/touch/paintevent/loadPixel.cpp
@@ -39,8 +39,7 @@ void TabletCanvas::load(QPainter &painter,
 {
     const bool withPdf          = dataPoint.withPdf;
     const bool is_play          = likely((dataPoint.parent)) ? (dataPoint.parent->m_audioplayer->isPlay()) : false;
-
-    const int m_pos_ris         = (is_play) ? (dataPoint.parent->m_audioplayer->getPositionSecond()) : -1;
+    const qint64 m_pos_ris      = (is_play) ? (dataPoint.parent->m_audioplayer->getPositionSecond()) : -1;
     static int last_m_pos_ris   = -1;
 
     int lenPage                     = data->datatouch->lengthPage();
@@ -151,7 +150,6 @@ static void loadSheet(
     const int lenPage = doc.datatouch->lengthPage();
     const double zoom = doc.datatouch->getZoom();
     const stroke *__stroke;
-
     int counterPage, counterPoint, lenPoint;
 
     pressure_t pressure;
